misc/HtmlTestSuite/results/test-0908.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Event handlers processing algorithm: error events</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="https://html.spec.whatwg.org/multipage/#the-event-handler-processing-algorithm" rel="help"/>
<link href="mailto:d@domenic.me" rel="author" title="Domenic Denicola"/>

</head><body><div id="log"/>

<script>
"use strict";
setup({ allow_uncaught_exception: true });

promise_test(t =&gt; {
  document.onerror = t.step_func((...args) =&gt; {
    assert_equals(args.length, 1);
    return true;
  });

  const eventWatcher = new EventWatcher(t, document, "error");
  const promise = eventWatcher.wait_for("error").then(e =&gt; {
    assert_equals(e.defaultPrevented, false);
    assert_equals(e.message, "");
    assert_equals(e.filename, "");
    assert_equals(e.lineno, 0);
    assert_equals(e.colno, 0);
    assert_equals(e.error, undefined);
  });

  document.dispatchEvent(new ErrorEvent("error", { cancelable: true }));

  return promise;
}, "error event is normal (return true does not cancel; one arg) on Document, with a synthetic ErrorEvent");

test(() =&gt; {
  const e = new ErrorEvent("error");
  assert_equals(e.message, "");
  assert_equals(e.filename, "");
  assert_equals(e.lineno, 0);
  assert_equals(e.colno, 0);
  assert_equals(e.error, undefined);
}, "Initial values of ErrorEvent members")

test(() =&gt; {
  const e = new ErrorEvent("error", {error : null});
  assert_equals(e.error, null);
}, "error member can be set to null")

test(() =&gt; {
  const e = new ErrorEvent("error", {error : undefined});
  assert_equals(e.error, undefined);
}, "error member can be set to undefined")

test(() =&gt; {
  const e = new ErrorEvent("error", {error : "foo"});
  assert_equals(e.error, "foo");
}, "error member can be set to arbitrary")

</script>
</body></html>
